test: HeroUnits
