+++ src/schema/v2/homeView/__tests__/HomeViewSection.test.ts
@@ -484,68 +484,6 @@ describe("HomeViewSection", () => {
     })
   })
 
-  describe("HeroUnits", () => {
-    it("returns lists of hero units", async () => {
-      const query = gql`
-        {
-          homeView {
-            section(id: "home-view-section-hero-units") {
-              __typename
-
-              ... on HomeViewSectionHeroUnits {
-                heroUnitsConnection(first: 2) {
-                  edges {
-                    node {
-                      title
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      `
-
-      const mockHeroUnitsResponse = {
-        body: [
-          {
-            title: "Foundations Summer 2024",
-          },
-          {
-            title: "Foundations Prize Finalists",
-          },
-        ],
-        headers: { "x-total-count": 2 },
-      }
-
-      const context = {
-        heroUnitsLoader: jest.fn().mockReturnValue(mockHeroUnitsResponse),
-      }
-
-      const { homeView } = await runQuery(query, context)
-
-      expect(homeView.section).toMatchInlineSnapshot(`
-        Object {
-          "__typename": "HomeViewSectionHeroUnits",
-          "heroUnitsConnection": Object {
-            "edges": Array [
-              Object {
-                "node": Object {
-                  "title": "Foundations Summer 2024",
-                },
-              },
-              Object {
-                "node": Object {
-                  "title": "Foundations Prize Finalists",
-                },
-              },
-            ],
-          },
-        }
-      `)
-    })
-  })
-
   describe("HomeViewSectionMarketingCollections", () => {
     it("returns correct data", async () => {
       const query = gql`
