+++ src/schema/v2/homeView/sections/__tests__/HeroUnits.test.ts
@@ -1,0 +1,101 @@
+import gql from "lib/gql"
+import { runQuery } from "schema/v2/test/utils"
+
+describe("HeroUnits", () => {
+  it("returns the section's metadata", async () => {
+    const query = gql`
+      {
+        homeView {
+          section(id: "home-view-section-hero-units") {
+            __typename
+            internalID
+            contextModule
+            ownerType
+            component {
+              title
+              description
+              behaviors {
+                viewAll {
+                  buttonText
+                  href
+                  ownerType
+                }
+              }
+            }
+          }
+        }
+      }
+    `
+
+    const context = {}
+
+    const { homeView } = await runQuery(query, context)
+
+    expect(homeView.section).toMatchInlineSnapshot(`
+      Object {
+        "__typename": "HomeViewSectionHeroUnits",
+        "component": null,
+        "contextModule": "heroUnitsRail",
+        "internalID": "home-view-section-hero-units",
+        "ownerType": null,
+      }
+    `)
+  })
+
+  it("returns the section's connection data", async () => {
+    const query = gql`
+      {
+        homeView {
+          section(id: "home-view-section-hero-units") {
+            ... on HomeViewSectionHeroUnits {
+              heroUnitsConnection(first: 2) {
+                edges {
+                  node {
+                    title
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    `
+
+    const mockHeroUnitsResponse = {
+      body: [
+        {
+          title: "Foundations Summer 2024",
+        },
+        {
+          title: "Foundations Prize Finalists",
+        },
+      ],
+      headers: { "x-total-count": 2 },
+    }
+
+    const context = {
+      heroUnitsLoader: jest.fn().mockReturnValue(mockHeroUnitsResponse),
+    }
+
+    const { homeView } = await runQuery(query, context)
+
+    expect(homeView.section).toMatchInlineSnapshot(`
+      Object {
+        "heroUnitsConnection": Object {
+          "edges": Array [
+            Object {
+              "node": Object {
+                "title": "Foundations Summer 2024",
+              },
+            },
+            Object {
+              "node": Object {
+                "title": "Foundations Prize Finalists",
+              },
+            },
+          ],
+        },
+      }
+    `)
+  })
+})
